Initial commit to Scorecard page

+++ app/src/main/java/com/cricketscoringapp/criceasy/Database/DatabaseHelper.java
@@ -2285,7 +2285,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 int fives = cursor.getInt(cursor.getColumnIndex(COLUMN_FIVES));
                 int sixes = cursor.getInt(cursor.getColumnIndex(COLUMN_SIXES));
                 // Add the Batter object to the list
-                Batsman batsman = new Batsman(playerId, playerName, inningsId, score, ballsPlayed, zeroes, ones, twos,
+                Batsman batsman = new Batsman(playerId, playerName, (int) inningsId, score, ballsPlayed, zeroes, ones, twos,
                         threes, fours, fives, sixes);
                 batterList.add(batsman);
             } while (cursor.moveToNext());
